merged cells handled
if ANY merged cells are found, an error is through. this has to be corrected as per the todo notes in the file.
excelError has been changed a bit, allows errorTypes to the UI can decide what to do for specific errors.
excelWorkbook has a function to check if the data has been loaded correctly and throws error if not.
The UI has a basic object that can display error messages based of the thrown error type.